Renames playerManager to playerService because this is not really a manager but a service
